+++ demos/react-native-supabase-group-chat/src/components/list/index.tsx
@@ -1,5 +1,6 @@
 import { FlashList, ListRenderItem } from '@shopify/flash-list';
 import { Separator, Text } from 'tamagui';
+import { type JSX } from 'react';
 
 export function List<T>({
   data,
@@ -29,7 +30,6 @@ export function List<T>({
       data={data}
       extraData={extraData}
       renderItem={renderItem}
-      estimatedItemSize={87}
       ItemSeparatorComponent={() => <Separator />}
       numColumns={numColumns}
       /* refreshControl={