+++ demos/react-native-supabase-group-chat/src/lib/date.ts
@@ -1,6 +1,0 @@
-import { formatRelative } from "date-fns";
-
-export function stringToRelativeDate(date: string) {
-  const dateObject = new Date(date);
-  return dateObject ? formatRelative(new Date(date), new Date()) : "";
-}