+++ packages/node/download_core.js
@@ -5,18 +5,18 @@ import { Readable } from 'node:stream';
 import { finished } from 'node:stream/promises';
 
 // When changing this version, run node download_core.js update_hashes
-const version = '0.4.6';
+const version = '0.4.10';
 const versionHashes = {
-  'powersync_x64.dll': '5efaa9ad4975094912a36843cb7b503376cacd233d21ae0956f0f4b42dcb457b',
-  'powersync_x86.dll': '4151ba8aa6f024b50b7aebe52ba59f2c5be54e3fed26f7f3f48e1127dcda027d',
-  'powersync_aarch64.dll': '3abe46074432593ff5cfc2098b186c592f020c5cfa81285f8e49962732a94bf5',
-  'libpowersync_x86.so': '1321a7de13fda0b2de7d2bc231a68cb5691f84010f3858e5cf02e47f88ba6f4a',
-  'libpowersync_x64.so': 'e9d78620d69d3cf7d57353891fe0bf85b79d326b42c4669b9500b9e610388f76',
-  'libpowersync_aarch64.so': '0d84c0dc0134fc89af65724d11e2c45e3c15569c575ecda52d0ec2fa2aeec495',
-  'libpowersync_armv7.so': 'c7887181ce9c524b68a7ac284ab447b8584511c87527ca26186e5874bf9ba3d6',
-  'libpowersync_riscv64gc.so': 'a89f3a71f22f707707d97517e9310e42e2a57dc5343cee08d09002a8cea048d5',
-  'libpowersync_x64.dylib': '9b484eaf361451f7758ca6ad53190a73563be930a8f8a39ccefd29390046ef6c',
-  'libpowersync_aarch64.dylib': 'bfb4f1ec207b298aff560f1825f8123d24316edaa27b6df3a17dd49466576b92'
+  'powersync_x64.dll': '9cffcd5e2393172523405c35021ce5db70f1d257a7973ca4210c67dbfd00c203',
+  'powersync_x86.dll': 'e809ca6c877bc011ce193441d922a9c9f7c8d43675c990c671c22983263aad41',
+  'powersync_aarch64.dll': '756c4b591562b468c4f7c4a2a6a9a910b52389760294bac0e02f1c25636e68cc',
+  'libpowersync_x86.linux.so': 'e0a4bf3e1228386fbdffafe4d8bc7d00b2e5d1f6340ec8db563a205bd3299121',
+  'libpowersync_x64.linux.so': '81ab16086e504cafc248969c44b53731bea937ea721b2f8cda78bcdc835de1f2',
+  'libpowersync_aarch64.linux.so': 'a2e87871ce8fd9418c6a4e6d29ba5b6024062f604dd48dc878022fd2be6884a3',
+  'libpowersync_armv7.linux.so': '857c673912db666ec4248f2fd5a942b8eec3c8a1830db2ece0dd8de171943f88',
+  'libpowersync_riscv64gc.linux.so': '45bf10394010f410f6ddb16299bbf326af1286b9146972b0c8e0aaa545aa9ddf',
+  'libpowersync_x64.macos.dylib': '21dedb45427e6abcc2dd080db82b4bf57a90ddc7ef811c87927a02dcd52188ac',
+  'libpowersync_aarch64.macos.dylib': 'cf8c2f6c3bb6ed18e58b415423d51db2ffbaadcc077cf3522ed7402ca56313ce'
 };
 
 const assets = Object.keys(versionHashes);