+++ tools/diagnostics-app/CHANGELOG.md
@@ -1,5 +1,23 @@
 # diagnostics-app
 
+## 0.9.21
+
+### Patch Changes
+
+- Updated dependencies [d0c67b1]
+- Updated dependencies [133f376]
+  - @powersync/web@1.31.0
+  - @powersync/react@1.8.2
+
+## 0.9.20
+
+### Patch Changes
+
+- Updated dependencies [299c6dc]
+- Updated dependencies [616c2a1]
+  - @powersync/web@1.30.0
+  - @powersync/react@1.8.2
+
 ## 0.9.19
 
 ### Patch Changes